altering code to create full and partial split

+++ src/main/java/uk/gov/hmcts/reform/fact/data/api/os/OsLocationData.java
@@ -5,8 +5,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -15,6 +13,6 @@ public class OsLocationData {
 
     private String postcode;
     private String authorityName;
-    private BigDecimal latitude;
-    private BigDecimal longitude;
+    private double latitude;
+    private double longitude;
 }
